requirements.txt
cogapp
six
lxml
cssselect
